Move classes from GitUIPluginInterfaces.BuildServerIntegration to GitExtensions.Plugins.BuildServer.Core
