+++ Plugins/BuildServer/Core/BuildServerAdapterMetadataAttribute.cs
@@ -1,7 +1,8 @@
 using System;
 using System.ComponentModel.Composition;
+using GitUIPluginInterfaces.BuildServerIntegration;
 
-namespace GitUIPluginInterfaces.BuildServerIntegration
+namespace GitExtensions.Plugins.BuildServer.Core
 {
     [MetadataAttribute]
     [AttributeUsage(AttributeTargets.Class)]
