+++ Plugins/BuildServerIntegration/AppVeyorIntegration/Settings/AppVeyorSettingsUserControl.cs
@@ -1,8 +1,8 @@
 using System.Collections.Generic;
 using System.ComponentModel.Composition;
 using System.Windows.Forms;
+using GitExtensions.Plugins.BuildServer.Core;
 using GitUIPluginInterfaces;
-using GitUIPluginInterfaces.BuildServerIntegration;
 using ResourceManager;
 
 namespace AppVeyorIntegration.Settings
