+++ Plugins/BuildServerIntegration/JenkinsIntegration/JenkinsAdapter.cs
@@ -14,6 +14,7 @@ using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
 using GitCommands.Utils;
+using GitExtensions.Plugins.BuildServer.Core;
 using GitUI;
 using GitUIPluginInterfaces;
 using GitUIPluginInterfaces.BuildServerIntegration;
