+++ Plugins/BuildServerIntegration/TeamCityIntegration/TeamCityIntegration.csproj
@@ -14,6 +14,7 @@
     <ProjectReference Include="..\..\..\GitCommands\GitCommands.csproj" />
     <ProjectReference Include="..\..\..\GitExtUtils\GitExtUtils.csproj" />
     <ProjectReference Include="..\..\..\ResourceManager\ResourceManager.csproj" />
+    <ProjectReference Include="..\..\BuildServer\GitExtensions.Plugins.BuildServer.csproj" />
     <ProjectReference Include="..\..\GitUIPluginInterfaces\GitUIPluginInterfaces.csproj" />
   </ItemGroup>
