+++ Plugins/BuildServerIntegration/TfsIntegration/Settings/TfsSettingsUserControl.cs
@@ -2,8 +2,8 @@ using System;
 using System.Collections.Generic;
 using System.ComponentModel.Composition;
 using System.Windows.Forms;
+using GitExtensions.Plugins.BuildServer.Core;
 using GitUIPluginInterfaces;
-using GitUIPluginInterfaces.BuildServerIntegration;
 using ResourceManager;
 
 namespace TfsIntegration.Settings
